DP4-1 - fix bugs, change spec and added haml
