--- app/controllers/pages_controller.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class PagesController < ApplicationController
-  def home; end
-end
