+++ app/controllers/static_pages_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class StaticPagesController < ApplicationController
+  def home; end
+
+  def admin_home; end
+end
