--- app/views/admin/index.html.erb
@@ -1,1 +1,0 @@
-<h1>Welcome ADMIN to app MY SHOP</h1>
