+++ app/views/static_pages/admin_home.html.haml
@@ -1,0 +1,1 @@
+%h1 Welcome ADMIN to app MY SHOP
