+++ app/views/static_pages/home.html.haml
