--- spec/routing/admin_spec.rb
@@ -1,9 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe 'Admin', type: :routing do 
-  describe 'GET /admin' do
-    it 'routes /admin to the /index action' do      
-      expect(get '/admin').to route_to('admin#index')
-    end
-  end
-end
